+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- `require_all_specified_bands` flag in config.yaml to specify if all specified bands need to be available for a tile to download it.
+- download verification by checking if downloaded files have the expected size
+- decompress compressed g-band and z-band files to speed up subsequent file loading times
+
 ## [0.1.1] - 2025-10-31
 
 ### Fixed